distance sensor implemented and working

+++ DistanceSensor.h
@@ -7,14 +7,14 @@ class DistanceSensor
 private:
 	const byte m_ultrasonicTrigPin;
 	const byte m_ultrasonicEchoPin;
-	const byte m_distanceThreshold;
+	const unsigned int m_distanceThreshold;
 	bool m_isAboveThreshold;
-	int m_distanceValue;
+	unsigned int m_distanceValue;
 
-	byte ReadDistance();
+	unsigned int ReadDistance();
 public:
 	DistanceSensor(byte, byte, byte);
 	bool GetDistanceStatus();
-	int GetDistanceValue();
+	unsigned int GetDistanceValue();
 };
